src/requirements.txt
boto3==1.26.47
celery==5.2.7
channels==4.0.0
channels-redis==4.0.0
coverage==7.0.5
daphne==4.0.0
Django==4.0.8
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-cprofile-middleware==1.0.5
django-debug-toolbar==3.8.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
fpdf2==2.6.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
pypdf2==3.0.1
redis==4.4.2
requests==2.28.1
sentry-sdk==1.17.0
